backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.2
pydantic-settings==2.3.4
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
redis==5.0.4
rq==1.16.2
boto3==1.34.129
pdfminer.six==20231228
python-docx==1.1.2
pytest==8.2.2
httpx==0.27.0
